requirements.txt
fastapi==0.75.0
google_search_results==2.4.1
googletrans==3.1.0a0
pydantic==1.8.2
requests==2.26.0
Unidecode==1.3.2
uvicorn
scikit-learn
nltk
Pillow
tqdm
numpy
gdown
transformers 
scikit-image
git+https://github.com/openai/CLIP.git
pytorch-pretrained-bert
ipython